refactor(osn/api): split the auth route monolith into route groups
routes/auth.ts (~1,800 lines, one chained Elysia app) is now a
routes/auth/ directory, mirroring the services/auth/ split:

- index.ts keeps the original createAuthRoutes signature (with its
  parameter docs), builds a shared AuthRouteContext once, and mounts
  one Elysia route group per domain via .use(). AuthRateLimiters +
  createDefaultAuthRateLimiters re-export from limiters.ts, so all
  existing "routes/auth" imports resolve unchanged.
- context.ts carries the shared plumbing: auth service instance, app
  runner, fail-closed rateLimit + turnstileGate helpers, IP resolution
  (S-M34), session-meta extraction, passkey-enroll principal
  resolution, and the prebuilt JWKS document.
- 13 route groups (registration, tokens, passkey-enroll, login,
  profile-switch, recovery, step-up, sessions, email-change,
  security-events, passkey-management, cross-device, well-known);
  handler bodies moved verbatim.

Also closes the T-S1 test gap from the PR review: direct coverage for
verifyStepUpForAccountDelete (S-C1 purpose binding, both rejection
directions) and verifyStepUpForExternalPurpose (accountId passthrough
+ purpose mismatch).

714 tests green; wiki path references updated and the route-layer
module-directory pattern noted in backend-patterns.
